Leveller: Changed type of class member from size_t to int, to fix problem with driver only working on 32-bit platforms. (#12166)

+++ frmts/leveller/levellerdataset.cpp
@@ -367,7 +367,7 @@ class digital_axis
 
   protected:
     int m_eStyle;
-    size_t m_fixedEnd;
+    int m_fixedEnd;
     double m_d[2];
 };
